refactor(payments): lift markRentalAsPaid into shared lib/payments namespace
Generalizes the email-claim dedup helper. Options shape changed:
{ transactionSn } -> { transactionRef } so Opn can pass chrg_* and
ShopeePay can pass transaction_sn through the same param. All
ShopeePay call sites updated in this commit. Re-export shim keeps
import paths stable.
